@@ -1,6 +1,6 @@
 'use strict';
 
-/* eslint-disable global-require */
+ 
 
 var fs = require('fs');
 var path = require('path');
