@@ -68,7 +68,7 @@ function warnForDeprecatedBehavior(options) {
 
     /* istanbul ignore if  */
     if (config.oldBehavior) {
-        // eslint-disable-next-line
+         
         console.warn('The rule `angular/service-name` will be split up to different rules in the next version. Please read the docs for more information');
     }
 }
