@@ -96,12 +96,12 @@ eslintTester.run('di-unused', rule, {
         },
         {
             code: 'angular.module("").controller("", ["$q", ($q) => {}]);',
-            parserOptions: {ecmaVersion: 6},
+            languageOptions: {ecmaVersion: 6},
             errors: [{message: 'Unused injected value $q'}]
         },
         {
             code: 'angular.module("").controller("", ($q) => {});',
-            parserOptions: {ecmaVersion: 6},
+            languageOptions: {ecmaVersion: 6},
             errors: [{message: 'Unused injected value $q'}]
         },
         // directive
@@ -133,14 +133,14 @@ eslintTester.run('di-unused', rule, {
         {
             code: 'angular.module("").factory("", $q => {});',
             errors: [{message: 'Unused injected value $q'}],
-            parserOptions: {ecmaVersion: 6}
+            languageOptions: {ecmaVersion: 6}
         }, {
             code: 'angular.module("").factory("", ["q", function($q) {}]);',
             errors: [{message: 'Unused injected value $q'}]
         }, {
             code: 'angular.module("").factory("", ["q", $q => {}]);',
             errors: [{message: 'Unused injected value $q'}],
-            parserOptions: {ecmaVersion: 6}
+            languageOptions: {ecmaVersion: 6}
         }, {
             code: 'var app = angular.module(""); app.factory("", function($q) {});',
             errors: [{message: 'Unused injected value $q'}]
