@@ -27,7 +27,7 @@ valid.push({
 }, {
     code: 'app.factory("eslintFactory", EslintFactory); class EslintFactory {}',
     options: ['eslint'],
-    parserOptions: {ecmaVersion: 6}
+    languageOptions: {ecmaVersion: 6}
 }, {
     code: 'app.factory("eslintFactory", function() {});',
     options: [/^eslint/]
